Fix entries/treatments JSON parsing crash on real Nightscout data
The diagnostic log pinned this exactly:

  JsonDecodingException: Unexpected symbol '.' in numeric literal
  at path: \$.result[0].date

Nightscout's real (non-empty) responses report `date` with a
fractional-millisecond component depending on the uploader, but the
DTOs declared it as Long, which kotlinx.serialization refuses to
coerce from a JSON float. Only showed up once there was live data to
fetch — the earlier empty-database testing couldn't have caught it.
Changed EntryDto.date and TreatmentDto.date to Double, truncating to
Long (irrelevant sub-millisecond precision loss) when mapping into the
Room entities.

+++ app/src/main/kotlin/com/trionsandroid/app/data/remote/EntryDto.kt
@@ -7,7 +7,9 @@ import kotlinx.serialization.Serializable
 data class EntryDto(
     @SerialName("_id") val legacyId: String? = null,
     val identifier: String? = null,
-    val date: Long,
+    // Nightscout sometimes reports `date` with a fractional-millisecond component
+    // (e.g. 1787474479807.234) depending on the uploader, so this can't be a Long.
+    val date: Double,
     val sgv: Int? = null,
     val direction: String? = null,
     val type: String? = null,
